refactored health checking into its own function that cli and endpoint calls
