+++ tests/utils/test_check_clickhouse.py
@@ -6,11 +6,11 @@ from unittest import mock
 
 import pytest
 
-from snuba.cli.health import check_clickhouse, filter_checked_storages
 from snuba.datasets.dataset import Dataset
 from snuba.datasets.factory import get_dataset
 from snuba.datasets.readiness_state import ReadinessState
 from snuba.datasets.storage import ReadableTableStorage
+from snuba.utils.health_info import check_clickhouse, filter_checked_storages
 
 
 class BadStorage(mock.MagicMock):
@@ -88,20 +88,20 @@ def temp_settings() -> Any:
 
 
 @mock.patch(
-    "snuba.web.views.get_enabled_dataset_names",
+    "snuba.utils.health_info.get_enabled_dataset_names",
     return_value=["events"],
 )
-@mock.patch("snuba.web.views.get_dataset", side_effect=fake_get_dataset)
+@mock.patch("snuba.utils.health_info.get_dataset", side_effect=fake_get_dataset)
 @pytest.mark.clickhouse_db
 def test_check_clickhouse(mock1: mock.MagicMock, mock2: mock.MagicMock) -> None:
     assert check_clickhouse()
 
 
 @mock.patch(
-    "snuba.web.views.get_enabled_dataset_names",
+    "snuba.utils.health_info.get_enabled_dataset_names",
     return_value=["events", "bad"],
 )
-@mock.patch("snuba.web.views.get_dataset", side_effect=fake_get_dataset)
+@mock.patch("snuba.utils.health_info.get_dataset", side_effect=fake_get_dataset)
 def test_bad_dataset_fails_healthcheck(
     mock1: mock.MagicMock, mock2: mock.MagicMock
 ) -> None:
@@ -111,7 +111,7 @@ def test_bad_dataset_fails_healthcheck(
 
 
 @mock.patch(
-    "snuba.web.views.get_enabled_dataset_names",
+    "snuba.utils.health_info.get_enabled_dataset_names",
     return_value=["events"],
 )
 @mock.patch("snuba.clusters.cluster._get_storage_set_cluster_map", return_value={})
@@ -125,10 +125,10 @@ def test_dataset_undefined_storage_set(
 
 
 @mock.patch(
-    "snuba.web.views.get_enabled_dataset_names",
+    "snuba.utils.health_info.get_enabled_dataset_names",
     return_value=["events", "experimental", "mock"],
 )
-@mock.patch("snuba.web.views.get_dataset", side_effect=fake_get_dataset)
+@mock.patch("snuba.utils.health_info.get_dataset", side_effect=fake_get_dataset)
 @pytest.mark.clickhouse_db
 def test_filter_checked_storages(
     mock1: mock.MagicMock, mock2: mock.MagicMock, temp_settings: Any
