+++ tests/web/test_views.py
@@ -82,3 +82,26 @@ def test_check_envoy_health(snuba_api: FlaskClient) -> None:
     with mock.patch("snuba.web.views.check_down_file_exists", return_value=True):
         response = snuba_api.get("/health_envoy")
         assert response.status_code == 503
+
+
+def test_check_health(snuba_api: FlaskClient) -> None:
+    response = snuba_api.get("/health")
+    assert response.status_code == 200
+    # down file existing does not mean the pod is unhealthy
+    with mock.patch(
+        "snuba.utils.health_info.check_down_file_exists", return_value=True
+    ):
+        response = snuba_api.get("/health")
+        assert response.status_code == 200
+    # don't check clickhouse if not thorough
+    with mock.patch("snuba.utils.health_info.check_clickhouse", return_value=False):
+        response = snuba_api.get("/health")
+        assert response.status_code == 200
+    # thorough healthcheck fails on bad clickhouse connection
+    with mock.patch("snuba.utils.health_info.check_clickhouse", return_value=False):
+        response = snuba_api.get("/health?thorough=true")
+        assert response.status_code == 502
+    # thorough healthcheck passes on good clickhouse connection
+    with mock.patch("snuba.utils.health_info.check_clickhouse", return_value=True):
+        response = snuba_api.get("/health?thorough=true")
+        assert response.status_code == 200
